Add first version of nice_agent_set_port_exclusions()
Sets a list of ports that should not be used for allocating host
candidates.

The provided ports is a comma-separated list of "port definitions". A
"port definition" can be a single port number (e.g. "12345"), or a port
range (e.g. "11111-22222").

Example input: "1111,2222,5000-6000"
This would avoid using ports 1111, 2222, and all ports between 5000 and
6000 (inclusive)